requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
binaryornot==0.4.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.4
et_xmlfile==2.0.0
fastapi==0.128.1
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.4.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==23.2
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.3.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==7.4.2
repath==0.9.0
reportlab==4.4.9
requests==2.32.5
rich==14.3.2
six==1.17.0
starlette==0.50.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==4.0.2
watchfiles==1.1.1
websockets==16.0
